Update models, repository, day query

+++ api.schedule/schedule-main-service/internal/model/group.go
@@ -1,10 +1,10 @@
 package model
 
 type Group struct {
-	ID             uint64
+	ID             int64
 	Name           string
 	Number         uint16
 	EnrollmentYear uint16
-	SpecialityID   uint64
-	CurrentTermID  uint8
+	SpecialityID   int64
+	CurrentTermID  int64
 }
