+++ api.schedule/schedule-main-service/internal/model/template.go
@@ -1,9 +1,9 @@
 package model
 
 type Template struct {
-	ID         uint64
-	GroupID    uint64
-	WeekTypeID uint8
-	DayID      uint8
-	TermID     uint8
+	ID         int64
+	GroupID    int64
+	WeekTypeID int64
+	DayID      int64
+	TermID     int64
 }
